+++ NEWS.md
@@ -19,7 +19,7 @@ Minor release.
 ## New features
 
 *  Added a `preamble` option for Slurm batch scripts. This allows the user to
-   specify commands that need to be added to script, e.g., `module load`.
+   specify commands that need to be added to the script, e.g., `module load`.
   
 *  `preamble` can be specified via `opts_slurmR$set_preamble()` or directly
    when calling `Slurm_*apply`.
@@ -38,7 +38,7 @@ Minor release.
 
 ## Bug fixes
 
-*  Fixes an user-message issue (not crucial) observed in Fedora and Solaris.
+*  Fixes a user-message issue (not crucial) observed in Fedora and Solaris.
 
 # slurmR 0.4-1 (CRAN)