+++ NEWS.md
@@ -6,10 +6,13 @@
 
 *  The function `sacct()` now doesn't fail when accounting is not available.
 
-*  slurmR now has a Docker image that can be used to test with a Slurm cluster.
+*  slurmR now has a Dockerimage that can be used to test with a Slurm cluster.
    The docker image can be started from within R (if Docker is available).
 
 *  `makeSlurmCluster()`s default tmp_path is now `opts_slurmR$get_tmp_path()`.
+
+*  New functions `slurmr_docker_*` provide a wrapper to interact with the
+   Dockerimage of slurmR.
 
 
 # slurmR 0.4-3 (CRAN)